PI-446 added tests for shortcuts
