+++ stackdio/core/tests/__init__.py
@@ -17,3 +17,4 @@
 
 
 from core.tests.unit import *  # NOQA
+from core.tests.integration import *  # NOQA
